+++ perllib/FixMyStreet/Cobrand/FixMyStreet.pm
@@ -10,8 +10,6 @@ use constant COUNCIL_ID_BROMLEY => 2482;
 
 sub on_map_default_status { return 'open'; }
 
-sub enable_category_groups { 1 }
-
 # Special extra
 sub path_to_web_templates {
     my $self = shift;